main.py
#!/usr/bin/env python
import os
from typing import Dict, List, Optional, Any
from datetime import datetime, timedelta
from fastapi import FastAPI, HTTPException, Depends, Request, BackgroundTasks
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel, Field, validator
import requests
from bs4 import BeautifulSoup
from dotenv import load_dotenv
from langchain_openai import ChatOpenAI, OpenAIEmbeddings
from langchain.text_splitter import RecursiveCharacterTextSplitter
from langchain_community.vectorstores import Chroma
from langchain.memory import ConversationBufferWindowMemory
from langchain.chains import ConversationalRetrievalChain
from langchain.prompts import PromptTemplate
import logging
from redis import Redis
import json

# ---------------------
# Setup Logging
# ---------------------
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

# ---------------------
# Load Environment Variables
# ---------------------
load_dotenv()
os.environ["OPENAI_API_KEY"] = os.getenv("OPENAI_API_KEY", "your_default_openai_key_if_needed")

class Config:
    OPENAI_API_KEY = os.environ["OPENAI_API_KEY"]
    REDIS_URL = os.getenv("REDIS_URL", "redis://localhost:6379")
    CHROMA_PERSIST_DIR = "chroma_db"
    MODEL_NAME = "gpt-4-turbo-preview"
    TEMPERATURE = 0.7
    CHUNK_SIZE = 1000
    CHUNK_OVERLAP = 200
    MEMORY_WINDOW_SIZE = 10
    MAX_TOKENS = 4000
    CACHE_EXPIRE = 3600

# (Your KnowledgeManager, RedisConversationMemory, ConversationManager, API endpoints, etc. would follow here)

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
