requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
redis==4.5.4
beautifulsoup4==4.12.2
requests==2.28.2
aiohttp==3.8.1
langchain==0.0.142
chromadb==0.3.21
